calculos varios

+++ app/src/main/java/com/example/usuario/comidaapp/Entidades/MenuDia.java
@@ -6,13 +6,13 @@ package com.example.usuario.comidaapp.Entidades;
 
 public class MenuDia {
     String nom_menu_dia;
-    Double cant_menu_dia;
+    int cant_menu_dia;
     String fecha_menu_dia;
     public MenuDia(){
 
     }
 
-    public MenuDia(String nom_menu_dia, Double cant_menu_dia, String fecha_menu_dia) {
+    public MenuDia(String nom_menu_dia, int cant_menu_dia, String fecha_menu_dia) {
         this.nom_menu_dia = nom_menu_dia;
         this.cant_menu_dia = cant_menu_dia;
         this.fecha_menu_dia = fecha_menu_dia;
@@ -26,11 +26,11 @@ public class MenuDia {
         this.nom_menu_dia = nom_menu_dia;
     }
 
-    public Double getCant_menu_dia() {
+    public int getCant_menu_dia() {
         return cant_menu_dia;
     }
 
-    public void setCant_menu_dia(Double cant_menu_dia) {
+    public void setCant_menu_dia(int cant_menu_dia) {
         this.cant_menu_dia = cant_menu_dia;
     }
